pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentpass-identity"
version = "0.2.1"
description = "AgentPass Python SDK for identity and access management"
readme = "README.md"
authors = [
    { name = "AgentPass Team", email = "contact@agentpass.com" }
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "pyjwt>=2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
fastapi = [
    "starlette>=0.27.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0"
]

[project.urls]
Homepage = "https://agentpass.com"
Documentation = "https://docs.agentpass.com"
Repository = "https://github.com/stefall-code/agentpass-sdk"

[tool.hatch.build.targets.wheel]
packages = ["src/agentpass"]